+++ rsecc.h
@@ -26,6 +26,6 @@
 #ifndef RSECC_H
 #define RSECC_H
 
-extern int RSECC_encode(int data_length, int ecc_length, const unsigned char *data, unsigned char *ecc);
+extern int RSECC_encode(size_t data_length, size_t ecc_length, const unsigned char *data, unsigned char *ecc);
 
 #endif /* RSECC_H */
